Bug observed on MacOS running scikitlearn.svm.SVC solved
On MacOs, svm.SVC raised a “typeerror: a float is required” error for
gamma = ‘scale’ (which was not raised running the same code in Ubuntu).
No idea how/why that happens.

The documentation of scikitlearn
(https://scikit-learn.org/stable/modules/generated/sklearn.svm.SVC.html)
 specifies the following info for gamma values: “…if gamma='scale' is
passed then it uses 1 / (n_features * X.var()) as value of gamma…” The
error indicates that gamma should be a float. So my guess was that the
string ‘scale’ could not be converted to float as the documentation
explains. I simply replaced it with the operation that results in a
float